backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
python-dotenv==1.1.1
PyJWT==2.9.0
cryptography==43.0.3
stripe==10.6.0
